refactor dubbo-samples-attachment
