+++ 2-advanced/dubbo-samples-attachment/dubbo-samples-attachment-consumer/src/main/resources/log4j.properties
