+++ 2-advanced/dubbo-samples-attachment/dubbo-samples-attachment-consumer/src/test/java/org/apache/dubbo/samples/attachment/AttachmentServiceIT.java
@@ -1,0 +1,51 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.dubbo.samples.attachment;
+
+import org.apache.dubbo.rpc.RpcContext;
+import org.apache.dubbo.samples.attachment.api.AttachmentService;
+import org.apache.dubbo.spring.boot.autoconfigure.DubboAutoConfiguration;
+
+import org.junit.Assert;
+import org.junit.Test;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.runner.RunWith;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@SpringBootTest(classes = {DubboAutoConfiguration.class})
+@RunWith(SpringRunner.class)
+public class AttachmentServiceIT {
+    @Autowired
+    private AttachmentService service;
+
+    @Test
+    public void testAttachment() {
+        RpcContext.getContext().setAttachment("index", "1");
+        String result = service.sayHello("dubbo");
+        Assert.assertTrue(result.startsWith("Hello dubbo"));
+        Assert.assertTrue(result.endsWith("index: 1"));
+        RpcContext.getContext().setAttachment("index", "2");
+        result = service.sayHello("dubbo");
+        Assert.assertTrue(result.endsWith("index: 2"));
+        result = service.sayHello("dubbo");
+        Assert.assertTrue(result.endsWith("index: null"));
+    }
+
+
+}
