+++ 2-advanced/dubbo-samples-attachment/dubbo-samples-attachment-provider/src/main/resources/application.yml
@@ -14,24 +14,14 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-services:
-  attachment-provider:
-    type: app
-    basedir: dubbo-samples-attachment-provider
-    mainClass: org.apache.dubbo.samples.attachment.AttachmentProvider
-  attachment-consumer:
-    type: test
-    basedir: dubbo-samples-attachment-consumer
-    tests:
-      - "**/*IT.class"
-    systemProps:
-      - zookeeper.address=attachment-provider
-      - zookeeper.port=2181
-      - dubbo.address=attachment-provider
-      - dubbo.port=20880
-    waitPortsBeforeRun:
-      - attachment-provider:2181
-      - attachment-provider:20880
-    depends_on:
-      - attachment-provider
-
+dubbo:
+  application:
+    name: attachment-provider
+    qos-enable: false
+  protocol:
+    name: dubbo
+    port: 20880
+  registry:
+    address: zookeeper://${zookeeper.address:127.0.0.1}:2181
+  provider:
+    token: "true"
